Divided into files handlers and keyboards refactor: common_handlers.py solo_handlers.py team_handlers.py splited handlers refactor: reply_keyboards.py inline_keyboards.py splited keyboards feat: inline_keyboards.py add keyboard for verification
