+++ lexicon/lexicon.py
@@ -11,5 +11,7 @@ LEXICON = {
     "add_teammate_button"   : "Добавить игрока",
     "team_done_button"      : "Команда собрана",
     "yes_stop_button"       : "Да, я хочу прекратить регистрацию",
-    "no_stop_button"        : "Нет, я хочу продолжить регистрацию"
+    "no_stop_button"        : "Нет, я хочу продолжить регистрацию",
+    "yes_verify_btn"        : "Подтвердить регистрацию",
+    "no_verify_btn"         : "Отменить регистрацию"
 }
